Forcemetadata update; specify the "--forcemetadata" parameter to always copy the metadata from the local picture to the remote picture in Picasa Web Album. This can be used to forcefully update remote metadata if local metadata changes, without uploading the entire picture/video again.
Note: yes, we could first retrieve the remote metadata, then see if there are changes and only if that is the case, perform the update (which is unfortunately slow). At the moment, with only "title" and "summary" fields set, that could be a nice performance tweak; however, in the future when we hope to add people tagging, it might turn out not to be possible or too complicated to first do a retrieval (for people data of XMP data from Picasa), etc etc... So first comparing before updating might possibly be handled in a future update.